Refactor check_for_loop_arg; rename manual_flatten's lint back to check_manual_flatten
